+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 # Release Notes
 
 ## v1.11.2
+### New
+
+- Added an optional, customizable info tooltip when hovering a game cover in the Library. (suggested by Colosso)
+- Added Metacritic scores from Steam, available to sort and filter by. (suggested by Colosso)
+
 ### Fixes
 
 - Fixed game names sorting capitalised titles ahead of all lowercase ones instead of alphabetically. (reported by Mayanaise)